src/requirements.txt
mysql-connector-python==8.2.0
networkx==3.2.1
numpy==1.26.4
pandas==2.1.4
python_dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.11.4
